Move strcut to other class.
