+++ SynicSugar/Assets/SynicSugar/Runtime/RTC/RTCStruct.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 18ce15d37d1e1104ca4e65f3621764a3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
